add default namespace for xml builds
