+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace]
-members = ["parse", "coyote", "html", "template_string", "component_string", "coyote_html", "rulesets",]
+members = ["parse", "coyote", "html", "template_steps", "component_string", "coyote_html", "rulesets",]
 resolver = "2"
 
 [workspace.dependencies]
