+++ coyote_html/Cargo.toml
@@ -9,5 +9,5 @@ edition = "2021"
 rulesets = { path = "../rulesets" }
 html = { path = "../html" }
 coyote = { path = "../coyote" }
-template_string = { path = "../template_string" }
+template_steps = { path = "../template_steps" }
 component_string = { path = "../component_string" }
