+++ template_steps/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "template_string"
+name = "template_steps"
 version = "0.1.0"
 edition = "2021"
